Feature/read totals also from tim (#61)
* +ADD: Calculate the totals List of a given TimeLogFile
-FIX: DeviceElements can have an id that only fits in an integer
-FIX: Change DDI from int to ushort
=OPT: Add funcion to create IntegerId from String
=UPD: Update Version

* Fix Version in ReleaseNotes

* Add DataLogTrigger only, if none is available so far

* +ADD: Get all TotalDDIs
+ADD: Get Type of DeviceProcessData
+ADD: Generate TIM-Element with all Totals

* +ADD: Function Description ISODevice.Analyse

* +ADD: In case no total was found in TimeLogs, the TIM-Element is taken into account

* +ADD: Make Checking of TIMElements optional with additional parameter

* +UPD: Update Version File

* +ADD: FUnction Descriptions

* -FIX: Hand GPSOptions instead of DeviceList to the TLG-Generator.

* Change DDI-VariableType to ushort for remaining functions
+ADD: TestCase für DDIs, die nicht im TimeLog zu finden waren

* +Add: Generate TotalsList
ATTENTION: Remaining Bug: No Lifetime Totals!

* -FIX: LifeTime Values were overwritten by ordinary Totals.
+ADD: Function to check if a specific DDI is a Total or LifeTimeTotal in a DeviceDescription

* +ADD: Function Description headers
-DEL: Delete unused usings

* +ADD: Function descriptions

* +ADD: TestCase for Totals-Generation

+++ Dev4Agriculture.ISO11783.ISOXML/TimeLog/ISOTLGExtract.cs
@@ -85,12 +85,12 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TimeLog
 
     public class ISOTLGExtract
     {
-        public readonly int Ddi;
+        public readonly ushort Ddi;
         public readonly int Det;
         public readonly string Name;
         public List<ISOTLGExtractPoint> Data { get; private set; }
 
-        public ISOTLGExtract(int ddi, int det, string name, List<ISOTLGExtractPoint> data)
+        public ISOTLGExtract(ushort ddi, int det, string name, List<ISOTLGExtractPoint> data)
         {
             Ddi = ddi;
             Det = det;
@@ -110,7 +110,7 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TimeLog
         /// <param name="lastValue"> Default set to "NO VALUE": The last known Value. This should *only* be used if the call for this function
         /// is part of loop for multiple TimeLogs. In case there are no values recorded for this TimeLog, this ensures that there are Entries created anyway</param>
         /// <returns></returns>
-        public static ISOTLGExtract FromTimeLog(ISOTLG timeLog, ushort ddi, short det = 0, string name = "", bool fillLines = false, int lastValue = ISOTLGExtractPoint.TLG_VALUE_FOR_NO_VALUE)
+        public static ISOTLGExtract FromTimeLog(ISOTLG timeLog, ushort ddi, int det = 0, string name = "", bool fillLines = false, int lastValue = ISOTLGExtractPoint.TLG_VALUE_FOR_NO_VALUE)
         {
 
             var entries = new List<ISOTLGExtractPoint>();
